refactor(components/atom/actionButton): add import file extensions
