+++ components/atom/actionButton/test/index.test.js
@@ -10,7 +10,7 @@ import ReactDOM from 'react-dom'
 import chai, {expect} from 'chai'
 import chaiDOM from 'chai-dom'
 
-import * as pkg from '../src'
+import * as pkg from '../src/index.js'
 
 import json from '../package.json'
